front (falta corregir Dropzone), spech2textHablantes

+++ controller.py
@@ -1,50 +1,53 @@
 # importing libraries
 import os
-from messages import messages
 from pathlib import Path
-from datetime import datetime
+
+import model
+import normalizar_audio
 import services.cloud_storage as cloud_storage
 import services.speech_to_text as speech_to_text
-import normalizar_audio
-import model
-from  post_procesar_transcripcion import deserialize_transcript
+import services.speech_to_text_dropSilences as s2t_dropSilence
+from messages import messages
+from post_procesar_transcripcion import deserialize_transcript
 
-def process_audio(file):
-    hourIni = datetime.now()
-    formato = "%H:%M:%S"
+
+formato = "%H:%M:%S"
+
+
+def process_audio(file, modeTrancript):
+    
     source_file = file
-    result=""
+    result = ""
+    
     if os.path.exists(source_file):
-        print('\33[32m' + hourIni.strftime(formato) + ' START MAIN' + '\033[0m')
-
         base = os.path.basename(source_file)
         name = os.path.splitext(base)[0]
         output_name = name + "_output.json"
+        
         try:
-
-
-            # --- Proceso Normalizar ---
-            # print('\33[32m' + datetime.now().strftime(formato) + '\033[0m')
-
+            # Proceso Normalizar
             normalizar_audio.normalizar(source_file)
+            # Proceso de Trancribir segun el modo de transcripcion
+            if modeTrancript == "hablantes":
+                storage_uri = cloud_storage.upload_blob(source_file, base)
+                result_json_array = speech_to_text.transcribe(storage_uri, Path(__file__).parent / "output" /output_name)
+                transcript = deserialize_transcript(result_json_array, True)
+            else:
+                transcript = s2t_dropSilence.transcribe(source_file)
 
-            # --- Procesos Transcribir ---
-            print('\33[32m' + datetime.now().strftime(formato) + '\033[0m')
-
-            storage_uri = cloud_storage.upload_blob(source_file, base)
-
-            result_json_array = speech_to_text.transcribe(storage_uri, Path(__file__).parent / "output" /output_name)
-            deserialized_text = deserialize_transcript(result_json_array, True)
-            summary = model.generate_summary(deserialized_text.replace("\n", ""))
-            result = {"summary": summary, "transcript": deserialized_text}
+            # Generar resumen
+            summary = model.generate_summary(transcript.replace("\n", ""))
+            print(transcript)
+            print(summary)
+            result = {"summary": summary, "transcript": transcript}
+            
         except Exception as e:
             print(messages.ERR_UNEXPECTED)
             print(e)
         finally:
-            cloud_storage.delete_blob(base)
+            if modeTrancript == "hablantes":
+                cloud_storage.delete_blob(base)
 
-        print('\33[32m' + datetime.now().strftime(formato) + ' FINISH MAIN' + '\033[0m')
-        print('\33[32m' + "Duracion --> " + str(datetime.now() - hourIni) + '\033[0m')
     else:
         print("El archivo de entrada no existe")
